Att LivroRepository

+++ src/main/java/com/SpringJpa/demo/service/AutorService.java
@@ -1,7 +1,9 @@
 package com.SpringJpa.demo.service;
 
+import com.SpringJpa.demo.exceptions.OperacaoNaoPermitidaException;
 import com.SpringJpa.demo.model.Autor;
 import com.SpringJpa.demo.repository.AutorRepository;
+import com.SpringJpa.demo.repository.LivroRepository;
 import com.SpringJpa.demo.validator.AutorValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,12 +17,13 @@ public class AutorService {
 
 
     private final AutorRepository autorRepository;
-
+    private final LivroRepository livroRepository;
     private final AutorValidator validator;
 
 
-    public AutorService(AutorRepository autorRepository, AutorValidator validator) {
+    public AutorService(AutorRepository autorRepository, AutorValidator validator,LivroRepository livroRepository) {
         this.autorRepository = autorRepository;
+        this.livroRepository = livroRepository;
         this.validator = validator;
     }
 
@@ -42,6 +45,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor){
+        if(possuiLivros(autor)){
+            throw new OperacaoNaoPermitidaException("Autor possui livros cadastrados");
+        }
         autorRepository.delete(autor);
     }
 
@@ -58,4 +64,7 @@ public class AutorService {
         return autorRepository.findAll();
     }
 
+    public boolean possuiLivros(Autor autor){
+        return livroRepository.existsByAutor(autor);
+    }
 }
